+++ src/Badge/index.ts
@@ -1,0 +1,1 @@
+export { Badge, BadgeProps, BadgeSize } from './Badge'